refactoring uiEnhancer
