+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "save-my-phind",
-  "version": "2.2.14",
+  "version": "2.2.16",
   "license": "RMD-C v1.0 License",
   "scripts": {
     "syncVersion": "node manage/syncVersions.js",
