+++ src/activeTab/triggers.js
@@ -1,13 +1,12 @@
 import {detectPageLoad, domainExportChecker, domainLoadChecker} from "./checker/domainChecker";
 import {launchExport} from "./scraper/scraper";
-import {improveUI} from "./uiEnhancer/uiEnhancer";
+import {uiEnhancer} from "./uiEnhancer/uiEnhancer";
 import {detectClicksModal} from "./utils/detectClicksModal";
 
 export async function actionExtensionIconClicked() {
   const domainPage = await domainExportChecker();
   if (domainPage === null) return;
   launchExport(domainPage);
-
   detectClicksModal();
 }
 
@@ -16,5 +15,5 @@ export async function actionPageLoaded() {
   if (domain === null) return;
   detectPageLoad(domain);
   // scrapOnLoadListener();
-  improveUI();
+  uiEnhancer(domain);
 }
