+++ src/activeTab/uiEnhancer/phind/exportButtons/createButtons.js
@@ -1,7 +1,7 @@
-import {waitAppears, waitAppend} from "../../interactDOM/insertElements";
+import {waitAppears, waitAppend} from "../../interactDOM/interactDOM";
 import {btnBarAllInline} from "./changeButtons";
 import {setBtnsExport} from "./styleExportButtons";
-import {createSideMenuBtn, createTopBtn} from "../generalElements/createElements";
+import {createSideMenuBtn, createTopBtn} from "../genericElements/createElements";
 import {launchExport} from "../../../scraper/scraper";
 
 async function setupButtons() {
